Change command structure
