+++ cli/internal/cmd/install/config.go
@@ -1,4 +1,4 @@
-package cmd
+package install
 
 import (
 	"context"
