+++ cli/internal/cmd/login.go
@@ -159,3 +159,16 @@ func newLoginStatusCommand() *cobra.Command {
 		},
 	}
 }
+
+func NewLogoutCommand() *cobra.Command {
+	return &cobra.Command{
+		Use:                   "logout",
+		Short:                 "Logout from a server",
+		Long:                  `Logout from a server.`,
+		DisableFlagsInUseLine: true,
+		Args:                  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return controlplane.Logout()
+		},
+	}
+}
